// src/types/resources/Procedure.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCanonical } from '../simpleTypes/Canonical';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TDateTime } from '../simpleTypes/DateTime';
import { TPeriod } from '../partials/Period';
import { TQuantity } from '../partials/Quantity';
import { TRange } from '../partials/Range';
import { TProcedurePerformer } from '../partials/ProcedurePerformer';
import { TAnnotation } from '../partials/Annotation';
import { TProcedureFocalDevice } from '../partials/ProcedureFocalDevice';

export type TProcedure = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  instantiatesCanonical?: TCanonical[];
  instantiatesUri?: TUri[];
  basedOn?: TReference[];
  partOf?: TReference[];
  status: string;
  statusReason?: TCodeableConcept;
  category?: TCodeableConcept;
  code?: TCodeableConcept;
  subject: TReference;
  encounter?: TReference;
  performedDateTime?: TDateTime;
  performedPeriod?: TPeriod;
  performedString?: string;
  performedAge?: TQuantity;
  performedRange?: TRange;
  recorder?: TReference;
  asserter?: TReference;
  performer?: TProcedurePerformer[];
  location?: TReference;
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  bodySite?: TCodeableConcept[];
  outcome?: TCodeableConcept;
  report?: TReference[];
  complication?: TCodeableConcept[];
  complicationDetail?: TReference[];
  followUp?: TCodeableConcept[];
  note?: TAnnotation[];
  focalDevice?: TProcedureFocalDevice[];
  usedReference?: TReference[];
  usedCode?: TCodeableConcept[];
};